Fix Directory/File not found

+++ EnvironmentServer.Daemon/Actions/CloneProductionTemplate.cs
@@ -34,8 +34,10 @@ public class CloneProductionTemplate : ActionBase
 
         await Bash.CommandAsync($"composer install -q", homeDir, validation: false);
 
-        File.Delete($"{homeDir}/vendor/shopware/recovery/composer.lock");
-        File.Delete($"{homeDir}/vendor/shopware/recovery/Common/composer.lock");
+        if (File.Exists($"{homeDir}/vendor/shopware/recovery/composer.lock"))
+            File.Delete($"{homeDir}/vendor/shopware/recovery/composer.lock");
+        if (File.Exists($"{homeDir}/vendor/shopware/recovery/Common/composer.lock"))
+            File.Delete($"{homeDir}/vendor/shopware/recovery/Common/composer.lock");
 
         if (File.Exists($"{homeDir}/public/.htaccess.dist"))
             File.Move($"{homeDir}/public/.htaccess.dist", $"{homeDir}/public/.htaccess");
